Package the repo using setup.py
