+++ core/scrapers/comidoc.py
@@ -2,7 +2,6 @@ import json
 import logging
 from typing import Dict, List
 
-import aiohttp
 from bs4 import BeautifulSoup
 
 from core.http import get
@@ -68,27 +67,33 @@ class ComidocScraper(BaseScraper):
         """
 
         text = await get(self.DOMAIN, headers=self.HEADERS)
-        soup = BeautifulSoup(text.decode("utf-8"), "html.parser")
+        if text is not None:
+            soup = BeautifulSoup(text.decode("utf-8"), "html.parser")
 
-        # We get the url hash needed from the path of the _buildManifest.js file
-        path_js = None
-        for i in soup.find_all("script"):
-            src = i.get("src", "")
-            if src.startswith("https://cdn.comidoc.net/_next/static/") and src.endswith(
-                "_buildManifest.js"
-            ):
-                path_js = src.split("/")[-2]
-                break
+            # We get the url hash needed from the path of the _buildManifest.js file
+            path_js = None
+            for i in soup.find_all("script"):
+                src = i.get("src", "")
+                if src.startswith(
+                    "https://cdn.comidoc.net/_next/static/"
+                ) and src.endswith("_buildManifest.js"):
+                    path_js = src.split("/")[-2]
+                    break
 
-        data = {}
-        # Fetch the daily courses if the path has been correctly resolved
-        if path_js is not None:
-            daily_json_link = f"{self.DOMAIN}/_next/data/{path_js}/daily.json"
-            data = await get(daily_json_link, headers=self.HEADERS)
+            data = {}
+            # Fetch the daily courses if the path has been correctly resolved
+            if path_js is not None:
+                daily_json_link = f"{self.DOMAIN}/_next/data/{path_js}/daily.json"
+                data = await get(daily_json_link, headers=self.HEADERS)
 
-            if data is not None:
-                data = json.loads(data)["pageProps"]["coupons"]
-            else:
-                data = {}
-                logger.warning(f"Empty response from comidoc. API may have changed!")
+                if data is not None:
+                    data = json.loads(data)["pageProps"]["coupons"]
+                else:
+                    data = {}
+                    logger.warning(
+                        f"Empty response from comidoc. API may have changed!"
+                    )
+        else:
+            data = {}
+            logger.warning("Error while fetching data from comidoc")
         return data
